refactor: make file names casing consistent
Make file names casing consistent with TypeScript conventions
