+++ src/providers/terminal-provider.ts
