+++ src/runners/mocha-test-runner.ts
@@ -1,12 +1,12 @@
 import {join} from 'path'
 import {debug, WorkspaceFolder} from 'vscode'
-import {ITestRunnerInterface} from '../interfaces/ITestRunnerInterface'
-import {ConfigurationProvider} from '../providers/ConfigurationProvider'
-import {TerminalProvider} from '../providers/TerminalProvider'
+import {ConfigurationProvider} from '../providers/configuration-provider'
+import {TerminalProvider} from '../providers/terminal-provider'
+import {TestRunner} from './test-runner'
 
 // TODO: Make a more generic test runner class and extend it
-export class AvaTestRunner implements ITestRunnerInterface {
-  public name = 'ava'
+export class MochaTestRunner implements TestRunner {
+  public name = 'mocha'
   public path: string = join('node_modules', '.bin', this.name)
   public terminalProvider: TerminalProvider
   public configurationProvider: ConfigurationProvider
@@ -28,9 +28,7 @@ export class AvaTestRunner implements ITestRunnerInterface {
     const additionalArguments = this.configurationProvider.additionalArguments
     const environmentVariables = this.configurationProvider.environmentVariables
 
-    const command = `${this.path} ${this.transformFileName(
-      fileName
-    )} -m "${testName}" ${additionalArguments}`
+    const command = `${this.path} ${fileName} --fgrep="${testName}" ${additionalArguments}`
 
     const terminal = this.terminalProvider.get({env: environmentVariables}, rootPath)
 
@@ -44,28 +42,14 @@ export class AvaTestRunner implements ITestRunnerInterface {
     const skipFiles = this.configurationProvider.skipFiles
 
     debug.startDebugging(rootPath, {
+      args: [fileName, '--fgrep', testName, ...additionalArguments.split(' ')],
       console: 'integratedTerminal',
       env: environmentVariables,
       name: 'Debug Test',
-      outputCapture: 'std',
-      port: 9229,
+      program: join(rootPath.uri.fsPath, this.path),
       request: 'launch',
-      runtimeArgs: [
-        'debug',
-        '--break',
-        '--serial',
-        this.transformFileName(fileName),
-        `--match="${testName}"`,
-        ...additionalArguments.split(' ')
-      ],
-      runtimeExecutable: join(rootPath.uri.fsPath, this.path),
       skipFiles,
       type: 'node'
     })
-  }
-
-  // We force slash instead of backslash for Windows
-  private transformFileName(fileName: string) {
-    return fileName.replace(/\\/g, '/')
   }
 }
