+++ src/runners/test-runner.ts
@@ -1,8 +1,8 @@
 import {WorkspaceFolder} from 'vscode'
-import {ConfigurationProvider} from '../providers/ConfigurationProvider'
-import {TerminalProvider} from '../providers/TerminalProvider'
+import {ConfigurationProvider} from '../providers/configuration-provider'
+import {TerminalProvider} from '../providers/terminal-provider'
 
-export interface ITestRunnerInterface {
+export interface TestRunner {
   name: string
   path: string
   terminalProvider: TerminalProvider
